+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/StandardRuleSetProvider.kt
@@ -11,6 +11,7 @@ public class StandardRuleSetProvider : RuleSetProvider {
         "standard",
         AnnotationRule(),
         AnnotationSpacingRule(),
+        ArgumentListWrappingRule(),
         ChainWrappingRule(),
         CommentSpacingRule(),
         FilenameRule(),
